Change test/report lifecycle
Tests used to run immediately upon invocation of `testcase` function.
The motivation for this was that it's easy to reason about, there is
no non-local execution of tests so we can setup our test fixtures
inline with the `testcase` in the root of a module/script.

But this had a couple of problems:

- If in future there is a mechanism to "skip" tests by filtering on
thier description, there would be no way to skip over setup/teardown
in the root scope.

- TAP output to stdout assumes sole ownership of the output stream,
as anything else may break the output. But if TAP tests are run from
many different modules at different times there is no well defined
scope for this ownership in immediate mode. If ducktest wants to e.g.
replace stdout for the duration of the output there is no nice way
to do this, the options are to preemptively wrap stdout as soon as
the ducktest module first runs (terrible!) or to wait until the first
test function is called (messy!).

- Responsibility for individual test output and responsibility for
"end of report" output (e.g. plan output, 1..n) are completely
separate from one another, making it hard to understand when and
where things happen. How do we know when all the tests have been
run? Who is responsible for making that determination?

For these reasons ducktest has been refactored so that `testcase`
merely *plans* a test. To execute the test and output a report
there is a new "report" function which optionally takes an output
stream and a reporter implementation as parameters.

This has a few positive effects;

- It better decouples reporting from testing, by not asking for a
reporter up-front before planning the test begins.

- It properly confines test output to a single (async) function
call, thus making it more reasonable to assert ownership of the
output stream.

- It gathers responsibility for report output to a single place.

By default when running in node, any tests planned for the
"default" suite (i.e. using the `testcase` `subcase` functions
exported from the module rather than creating new instances of
those functions by calling `suite` manually) a report is made to
stdout upon exit.

Further work:

Now `beforeAll` and `afterAll` style fixtures can't be set up
and torn down inline with calls to `testcase` in the root scope,
so we will ideally need a new `fixture` function to plan tests
which behave like this.