revert PR #9
Unfortunately this fix introduced a coupling between the default
output stream and the TAP format, by making the stream substitute
console log with a version that was aware of the TAP diagnostic
format.

The first shot at fixing this allowed the "write" property on
Stream to be optionally settable, then the tap reporter was made
responsible for making the substitution ... but thinking about it
it seems that there are other problems here:

- there are multiple ways to skin this cat; perhaps a
user would prefer to just completely suppress other log output?
Or to redirect it to a file? the tap reporter shouldn't make this
choice.

- the substitution is only made for the duration of the report
output, which still leaves the application free to throw out
whatever it wants before and after the report, which will still
break output.

- this still didn't deal with console.error, etc. or
process.stdout in node.

So all in all it was a bunch of extra complexity for a partial
solution. Now that the report lifecycle is already clearly managed
by the user, opening and closing a stream can be done at the call
site of `report` rather than fussing around with callbacks. And
wrapping a stream---which may require an entirely different
lifecycle---can be managaed elsewhere as appropriate for a particular
app/platform.