improved support for TAP format
- added optional plan for number of tests, and included plan output
for TAP report, automatically counting and putting at end if not
given. This fixes #17

- changed format for the diagnostic line that introduces a subtest.
Switched it from a diagnostic to `[description]` to make it more
consistent with the plan for concurrent subtests as per #1. This
is still forwards compatible with existing TAP parsers.

- added TAP version line at start of report output.

This commit also includes some refinement of Stream/Reporter/Report
API; now we don't always assume that a report must be associated
with an output stream. It's totally possible that this isn't always
the case!