Abstract out mechanism of synchronisation and state
This commit does not change any API, but it is a significant
rewrite of the sheduler. In particular, the way that
synchronisation is handled, and the way that state is coordinated
between enclosing and inner tests (in the absence of explicit
context passing to spec lambdas), is abstracted out.

(Note that naming of internal concepts is inconsistent and liable
to change, and in general the implementation needs cleaning up.
But those changes can be made incrementally on top of this, as
a first-cut of a working version.)

The motivation for this change is to support concurrent execution
of tests is node, as per #19, and there is more information there
about the platform differences between the web and node and why
the abstraction needs to be made.