Fixed issue with nested subtests
When descending through subtests we modify the subtest callback as
we go, to keep track of our progress. The mechanism by which this
is done is assigning to the local variable `encounterSubcase` in
the function `runSubcase`.

This variable was captured by the function returned from
`runSubcase` so that downstream callers see the change. But it
actually needed to be captured by the function *returned from the
function* returned from `runSubcase`.