Factor out a possibly-synchronous PromiseLike impl
We support both sycn and async tests. But if all our internal APIs
return normal promises then synchronous tests end up running async,
which is probably unexpected for users and is bad ergonomics. So
we have a few utility functions, `syncAsyncChain`, `syncAsyncCatch`,
etc. to operate on (T | Promise<T>), and behave synchronously when
possible.

This was messy and difficult to use, so instead we introduce a
PromiseLike class, Commitment, which exposes the same thing through
then, catch, finally. This makes code look more like idiomatic
Javascript, and makes everything easier to read.